+++ doc/src/history.md
@@ -7,7 +7,7 @@
 - `om ci`
   - Extra nix handling
       - Allow `--override-input` to work again (#439)
-      - Support `--rebuild` by disallowing it in irrelevant subcommands (`eval`, `flake lock`)
+      - Support `--rebuild` by disallowing it in irrelevant subcommands (`eval`, `develop`, `run`, `flake {lock,check}`, )
 
 ## 1.0.2 (2025-03-11) {#1.0.2}